Fix prefix handling.
Only an input or output action is allowed in prefix position.

Add a new action? predicate, which recognizes both input and
output actions, and change the contract for prefix so that
only an action is allowed in prefix position.

Change the parser to comply with this requirement; update most
of the test cases, as they expected incorrect processes to be
parsed correctly.