Handle repetitions in action object.
Using the same name twice as the object of an action is
an error: the new contract for input and output prevent it
from happening.

Fix some bits here and there to make sure all errors are catched
and processed correctly. Make all parsing routines raise exn:fail
instead of the too specific exn:fail:read.