Speed up collection availability check.
Instead of loading the whole library to check for its
availability, just make sure a suitable collection path exists
on the system.

Perhaps a little less reliable, but shaves a good couple seconds
off execution time for `make check', and any real problem with
the collection is going to pop up as soon as you try to run a
program using it anyway.